+++ src/os2datascanner/projects/report/reportapp/views/support_views.py
@@ -6,7 +6,7 @@ from ..utils import convert_context_to_email_body
 from ...organizations.models import Account
 from os2datascanner.core_organizational_structure.models.position import Role
 from os2datascanner.core_organizational_structure.models.organization import (
-    SupportContactChoices, DPOContactChoices)
+    SupportContactChoices, DPOContactChoices, MSGraphWritePermissionChoices)
 
 
 class SupportButtonView(TemplateView):
@@ -29,5 +29,7 @@ class SupportButtonView(TemplateView):
         context["organization"] = self.org
         context["SupportContactChoices"] = SupportContactChoices
         context["DPOContactChoices"] = DPOContactChoices
+        context["categorization"] = self.org.msgraph_write_permissions in (
+            MSGraphWritePermissionChoices.ALL, MSGraphWritePermissionChoices.CATEGORIZE)
 
         return context